use pci path for interface selection:
There was a change in systemd 239->240 where the vlan sub-interfaces
used to never match the link file, to always matching.  Vlan interfaces
share the same mac address so all vlan interfaces, on coming up, where
trying to be renamed by udev to the same name and would fail to come up.

Moving to PCI path fixes this.